flask-backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
google-cloud-dialogflowcx==1.24.0
